RDDL Support
 - Added helper class rddl.Task
