+++ src/tarski/rddl/__init__.py
@@ -1,0 +1,1 @@
+from .task import Task
